backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
gitpython
docker
httpx
pytest
